added by mohini nikhare

+++ Agriculture_Equipment_Loan_Application/src/main/java/com/financeCompany/agri/project/appDto/SanctionLetterDto.java
@@ -41,7 +41,7 @@ public class SanctionLetterDto {
 
 	private String rateOfInterest;
 
-	private double loanTenure;
+	private long loanTenure;
 
 	private double monthlyEmi;
